+++ sentry-android-core/src/main/java/io/sentry/android/core/SentryAndroidOptions.java
@@ -54,8 +54,8 @@ public final class SentryAndroidOptions extends SentryOptions {
    * Enables the Auto instrumentation for Activity lifecycle tracing.
    *
    * <ul>
-   *   <li>It also requires setting any of {@link SentryOptions#getEnableTracing()}, {@link
-   *       SentryOptions#getTracesSampleRate()} or {@link SentryOptions#getTracesSampler()}.
+   *   <li>It also requires setting any of {@link SentryOptions#getTracesSampleRate()} or {@link
+   *       SentryOptions#getTracesSampler()}.
    * </ul>
    *
    * <ul>