+++ sentry-spring-boot-jakarta/src/main/java/io/sentry/spring/boot/jakarta/SentryAutoConfiguration.java
@@ -454,10 +454,6 @@ public class SentryAutoConfiguration {
       super(ConfigurationPhase.REGISTER_BEAN);
     }
 
-    @ConditionalOnProperty(name = "sentry.enable-tracing")
-    @SuppressWarnings("UnusedNestedClass")
-    private static class SentryEnableTracingCondition {}
-
     @ConditionalOnProperty(name = "sentry.traces-sample-rate")
     @SuppressWarnings("UnusedNestedClass")
     private static class SentryTracesSampleRateCondition {}