+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -131,28 +131,8 @@ class SentryOptionsTest {
     }
 
     @Test
-    fun `when enableTracing is set to true tracing is considered enabled`() {
-        val options = SentryOptions().apply {
-            this.enableTracing = true
-        }
-
-        assertTrue(options.isTracingEnabled)
-    }
-
-    @Test
     fun `by default tracing is considered disabled`() {
         val options = SentryOptions()
-
-        assertFalse(options.isTracingEnabled)
-    }
-
-    @Test
-    fun `when enableTracing is set to false tracing is considered disabled`() {
-        val options = SentryOptions().apply {
-            this.enableTracing = false
-            this.tracesSampleRate = 1.0
-            this.tracesSampler = SentryOptions.TracesSamplerCallback { _ -> 1.0 }
-        }
 
         assertFalse(options.isTracingEnabled)
     }
@@ -354,7 +334,6 @@ class SentryOptionsTest {
         externalOptions.setTag("tag1", "value1")
         externalOptions.setTag("tag2", "value2")
         externalOptions.enableUncaughtExceptionHandler = false
-        externalOptions.enableTracing = true
         externalOptions.tracesSampleRate = 0.5
         externalOptions.profilesSampleRate = 0.5
         externalOptions.addInAppInclude("com.app")
@@ -399,7 +378,6 @@ class SentryOptionsTest {
         assertEquals(java.net.Proxy.Type.SOCKS, options.proxy!!.type)
         assertEquals(mapOf("tag1" to "value1", "tag2" to "value2"), options.tags)
         assertFalse(options.isEnableUncaughtExceptionHandler)
-        assertEquals(true, options.enableTracing)
         assertEquals(0.5, options.tracesSampleRate)
         assertEquals(0.5, options.profilesSampleRate)
         assertEquals(listOf("com.app"), options.inAppIncludes)